Add `PhysicsPlugins` plugin group and `PhysicsSetupPlugin`, refactor
- Removed `XpbdPlugin`

- Added `PhysicsPlugins` plugin group that initializes the plugins

- Added `setup` module and `PhysicsSetupPlugin` for
initializing schedules, sets and resources

- Renamed `steps` module to `plugins`

- Changed and added lots of docs
