+++ README.md
@@ -31,7 +31,7 @@ use bevy_xpbd_3d::prelude::*;
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
-        .add_plugin(XpbdPlugin)
+        .add_plugins(PhysicsPlugins)
         .add_startup_system(setup)
         .run();
 }
