+++ crates/examples_common_3d/src/lib.rs
@@ -1,19 +1,19 @@
 use bevy::{diagnostic::FrameTimeDiagnosticsPlugin, prelude::*};
 use bevy_inspector_egui::quick::WorldInspectorPlugin;
 
-use bevy_xpbd_3d::{XpbdLoop, XpbdPlugin};
+use bevy_xpbd_3d::prelude::*;
 
 #[derive(Default)]
 pub struct XpbdExamplePlugin;
 
 impl Plugin for XpbdExamplePlugin {
     fn build(&self, app: &mut App) {
-        app.add_plugin(XpbdPlugin)
+        app.add_plugins(PhysicsPlugins)
             .add_plugin(WorldInspectorPlugin::default())
             .add_plugin(FrameTimeDiagnosticsPlugin)
             .add_state::<AppState>()
-            .add_system(bevy_xpbd_3d::pause.in_schedule(OnEnter(AppState::Paused)))
-            .add_system(bevy_xpbd_3d::resume.in_schedule(OnExit(AppState::Paused)))
+            .add_system(pause.in_schedule(OnEnter(AppState::Paused)))
+            .add_system(resume.in_schedule(OnExit(AppState::Paused)))
             .add_system(pause_button)
             .add_system(step_button.run_if(in_state(AppState::Paused)));
     }
