+++ src/plugins/mod.rs
@@ -3,6 +3,7 @@
 pub mod broad_phase;
 pub mod integrator;
 pub mod prepare;
+pub mod setup;
 pub mod sleeping;
 pub mod solver;
 pub mod sync;
@@ -10,12 +11,13 @@ pub mod sync;
 pub use broad_phase::BroadPhasePlugin;
 pub use integrator::IntegratorPlugin;
 pub use prepare::PreparePlugin;
+pub use setup::*;
 pub use sleeping::SleepingPlugin;
 pub use solver::SolverPlugin;
 pub use sync::SyncPlugin;
 
 #[allow(unused_imports)]
-use crate::prelude::{steps::broad_phase::BroadCollisionPairs, *}; // For doc comments
+use crate::prelude::{plugins::broad_phase::BroadCollisionPairs, *}; // For doc comments
 use bevy::prelude::*;
 
 /// The main steps in the physics simulation loop.
